requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
diff-match-patch==20241021
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
django-import-export==4.3.7
django-oauth-toolkit==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
jwcrypto==1.5.6
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
